move signInWithGoogle to api.ts
